% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_net_sims.R
\name{combine_net_att_sims}
\alias{combine_net_att_sims}
\title{combine_net_att_sims}
\usage{
combine_net_att_sims(..., swapped_attribute)
}
\arguments{
\item{...}{Multiple returns of \code{run_att_swap}.}

\item{swapped_attribute}{Character string. Name of \code{attribute} given to \code{run_att_swap}.}
}
\value{
A datafarme with the network metrics and simulations results of all swap simulations.
}
\description{
to combine the various network simulaions into one dataframe for averaging
}
