% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_swap.R
\name{edge_swap}
\alias{edge_swap}
\title{Edge Swap}
\usage{
edge_swap(net_to_swap, swaps = 500, verbose = FALSE, edge_color = NULL)
}
\arguments{
\item{net_to_swap}{An object of type `network`. This is the network you want the swaps to be conducted on. All characteristics of this network, aside from the attribute you want to swap, will be the same in the result.}

\item{swaps}{Integer. The number of edge swaps you want to conduct. Defaults to 500.}

\item{verbose}{Logical. Would you like the function to display status messages? Defaults to `TRUE`.}

\item{edge_color}{Color as string or R color ID. Defaults to `NULL`.}
}
\value{
A network with edges randomly swapped `swaps` times with optional color assignments for the edges.
}
\description{
Edge Swap
}
