% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nattswap.R
\docType{package}
\name{nattswap}
\alias{nattswap}
\alias{nattswap-package}
\title{nattswap: A package for swapping network attributes to test if they are random.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{nattswap functions}{

The foo functions ...
}

