% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nattswap_means.R
\name{nattswap_means}
\alias{nattswap_means}
\title{nattswap_means}
\usage{
nattswap_means(nattswap_long)
}
\arguments{
\item{nattswap_long}{Output of \code{nattswap_transform}}
}
\value{
A table of means used for Z score tests.
}
\description{
nattswap_means
}
