% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nattswap_transform.R
\name{nattswap_transform}
\alias{nattswap_transform}
\title{nattswap_transform}
\usage{
nattswap_transform(comb_nattswaps)
}
\arguments{
\item{comb_nattswaps}{The object returned from \code{combine_net_att_sims}.}
}
\value{
A long format dataframe to be used for Z score testing.
}
\description{
To turn the combined nattswap from \code{combine_net_att_sims} from wide to long
}
