% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_swap.R
\name{net_att_swap}
\alias{net_att_swap}
\title{net_att_swap}
\usage{
net_att_swap(net_to_swap, attribute, swaps, verbose = FALSE)
}
\arguments{
\item{net_to_swap}{An object of type network. This is the network you want the swaps to be conducted on. All characteristics of this network, aside from the attribute you want to swap, will be the same in the result.}

\item{attribute}{A pre-defined attribute within net_to_swap. This is the attribute that already exists within the network object you want to randomize within the network. You must be able to call this attribute using net_to_swap\%v\%\\\"attribute\\\".}

\item{swaps}{Integer. The number of swaps between nodes you want to conduct.}

\item{verbose}{Logical. Would you like the function to display status messages? Defaults to FALSE.}
}
\value{
A network object with the attribute swapped a number of times equal to swaps.
}
\description{
To swap the attributes of a single network object
}
