% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_att_swap.R
\name{run_att_swap_sims}
\alias{run_att_swap_sims}
\title{run_att_swap_sims}
\usage{
run_att_swap_sims(
  net_to_swap,
  attribute,
  swaps = 10000,
  sims = 50,
  cores = 1L,
  save_preface = NA,
  save_dir = "./network_sims",
  verbose = FALSE
)
}
\arguments{
\item{net_to_swap}{An object of type \code{network}. Passed to \code{net_att_swap}.}

\item{attribute}{A pre-defined attribute within \code{net_to_swap}. Passed to \code{net_attr_swap}.}

\item{swaps}{Integer. Passed to net_attr_swap. Defaults to 10000.}

\item{sims}{How many simulated networks to generate. Each will be saved as an .rda file within save_dir. Defaults to 50.}

\item{cores}{Integer. How many cores to use. Defaults to 1.}

\item{save_preface}{Character. What to preface the above simulations with when saved. Defaults to sim.}

\item{save_dir}{A directory filepath. Where the individual network simulations and cumulative dataframe will be saved. Defaults to ./network_sims. Only used if save_preface is not NA.}

\item{verbose}{Logical. Should the code print more verbose progress messages. Defaults to FALSE.}
}
\value{
A list containing the original network with network metrics and all simulations.
}
\description{
run_att_swap_sims
}
