% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nattswap_compare.R
\name{nattswap_compare}
\alias{nattswap_compare}
\title{nattswap_compare}
\usage{
nattswap_compare(get_att_swap_sig_results, metric_name, swapped_att)
}
\arguments{
\item{get_att_swap_sig_results}{Result of \code{get_att_swap_sig}}

\item{metric_name}{Character. Either "degree", "evc", "n_bet", or "closeness"}

\item{swapped_att}{Character. Name of swapped attribute.}
}
\value{
A ggplot2 plot showing real vs simulated means
}
\description{
nattswap_compare
}
