% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nattswap_compare.R
\name{nattswap_compare}
\alias{nattswap_compare}
\title{nattswap_compare}
\usage{
nattswap_compare(
  get_att_swap_sig_results,
  metric_name = "all",
  swapped_att,
  all_nrow = 1,
  all_ncol = NULL,
  font_family = NULL,
  yaxis = c(NA, NA)
)
}
\arguments{
\item{get_att_swap_sig_results}{Result of \code{get_att_swap_sig}}

\item{metric_name}{Character. A vector of "degree", "evc", "bet", "n_bet", or "closeness". Defualts to "all" which will display all metrics.}

\item{swapped_att}{Character. Name of swapped attribute.}

\item{all_nrow}{How many rows do you want the plot to have?}

\item{all_ncol}{How many columns do you want the plot to have?}

\item{font_family}{Font family to use in plots.}

\item{yaxis}{Vector of 2 integers. Lower and upper limit of y axis.}
}
\value{
A ggplot2 plot showing real vs simulated means
}
\description{
nattswap_compare
}
